wip: move all tsv parsing code to the tsv package.
